🐛 fix : 쿠키 이름 변경 및 수정, 메뉴 수정
